pyproject.toml
[project]
name = "airflow-with-astro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.7",
    "apache-airflow-providers-common-io>=1.7.1",
    "apache-airflow-providers-git>=0.2.4",
    "apache-airflow-providers-openai>=1.7.2",
    "geopy>=2.4.1",
    "numpy>=2.4.2",
]
